move rectAddMargin to rectangle.js
