+++ src/fontra/client/core/rectangle.js
@@ -162,3 +162,12 @@ export function updateRect(rect, point) {
     yMax: Math.max(rect.yMax, point.y),
   };
 }
+
+export function rectAddMargin(rect, relativeMargin) {
+  const size = rectSize(rect);
+  const inset =
+    size.width > size.height
+      ? size.width * relativeMargin
+      : size.height * relativeMargin;
+  return insetRect(rect, -inset, -inset);
+}
