+++ src/fontra/views/editor/scene-controller.js
@@ -5,7 +5,7 @@ import { glyphLinesFromText, textFromGlyphLines } from "../core/glyph-lines.js";
 import { MouseTracker } from "../core/mouse-tracker.js";
 import { ObservableController } from "../core/observable-object.js";
 import { connectContours, splitPathAtPointIndices } from "../core/path-functions.js";
-import { insetRect, offsetRect, rectSize } from "../core/rectangle.js";
+import { insetRect, offsetRect, rectAddMargin, rectSize } from "../core/rectangle.js";
 import { packContour } from "../core/var-path.js";
 import { lenientIsEqualSet, isSuperset } from "../core/set-ops.js";
 import {
@@ -963,13 +963,4 @@ export function equalGlyphSelection(glyphSelectionA, glyphSelectionB) {
     glyphSelectionA?.lineIndex === glyphSelectionB?.lineIndex &&
     glyphSelectionA?.glyphIndex === glyphSelectionB?.glyphIndex
   );
-}
-
-export function rectAddMargin(rect, relativeMargin) {
-  const size = rectSize(rect);
-  const inset =
-    size.width > size.height
-      ? size.width * relativeMargin
-      : size.height * relativeMargin;
-  return insetRect(rect, -inset, -inset);
 }
